requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.32.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
yarl==1.7.2
